Update world-weather.py
I see that you have moved the get_user_location to the top, however in the meantime I merged a pull request from another user. We decided that it would be beneficial to get rid of gecoder module and use what's already there, so module request is used instead. He also worked on better error handling.